Use number row type in data types
Source: https://github.com/msupply-foundation/remote-server/commit/f5d431feac3f9aa6cef7aed69eec54d46fcdf2e9

+++ server/repository/src/schema/diesel_schema.rs
@@ -219,8 +219,9 @@ table! {
 table! {
     number (id) {
         id -> Text,
-        value -> BigInt,
+        value -> Integer,
         store_id -> Text,
+        #[sql_name = "type"] type_ -> crate::schema::number::NumberRowTypeMapping,
     }
 }
